refactor of blockFor overload

+++ vector-lerp/src/androidMain/kotlin/com/robgulley/time/Sleep.kt
@@ -5,7 +5,7 @@ import kotlin.time.ExperimentalTime
 
 actual object Sleep {
     actual fun blockFor(timeMillis: Long) = Thread.sleep(timeMillis)
-    actual fun blockFor(millis: Long, nanos: Int) = Thread.sleep(millis, nanos)
+    actual fun blockFor(millis: Long, nanos: Long) = Thread.sleep(millis, nanos.toInt())
 
     @OptIn(ExperimentalTime::class)
     actual fun blockFor(duration: Duration) {
